Combine transaction confirmation utils into one file
