+++ cardstack/src/hooks/transaction-confirmation/__tests__/extract-payload-params.test.ts
@@ -1,4 +1,4 @@
-import { extractPayloadParams } from '../extract-payload-params';
+import { extractPayloadParams } from '../utils';
 import { TypedData } from '@rainbow-me/model/wallet';
 
 describe('extractPayloadParams', () => {
